backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
passlib[argon2]==1.7.4
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.1.0
argon2-cffi==23.1.0
python-jose==3.3.0
python-multipart==0.0.6
alembic==1.12.1
google-generativeai==0.7.2